+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.2]
+
+- upgraded to `khao-malet@0.0.19`
+
+## [1.9.1]
+
+- upgraded to `khao-malet@0.0.18`
+- improve focus state for buttons
+- prevent empty hrefs for all buttons (to ensure focus)
+
 ## [1.9.0]
 
 - re-adjusted bg color for `khao-ui-card`